Added tests and retry on failure.
