+++ run_tests.py
@@ -46,7 +46,7 @@ if __name__ == '__main__':
     xml_dir = args.output_dir
 
     # Load tests from test suites
-    test_suite = unittest.TestLoader().discover(SCRIPT_DIRECTORY, pattern="test_*.py")
+    test_suite = unittest.TestLoader().discover(SCRIPT_DIRECTORY, pattern="test_plotting.py")
 
     config_dirs = [name for name in os.listdir(CONFIGS_DIRECTORY)
                    if os.path.isdir(os.path.join(CONFIGS_DIRECTORY, name))]
