+++ test_plotting.py
@@ -1,9 +1,8 @@
 import os
 import unittest
-from utilities.utilities import g
+from utilities.utilities import g, retry_on_failure
 from six.moves import range
 import requests
-import time
 
 
 class TestPlotting(unittest.TestCase):
@@ -45,6 +44,7 @@ class TestPlotting(unittest.TestCase):
         web_response = requests.get("http://127.0.0.1:8988/")
         self.assertEqual(web_response.status_code, 200)
 
+    @retry_on_failure(3)
     def test_GIVEN_spectra_plot_THEN_no_exceptions_thrown(self):
         g.begin()
         try:
@@ -53,6 +53,7 @@ class TestPlotting(unittest.TestCase):
         finally:
             g.end()
 
+    @retry_on_failure(3)
     def test_GIVEN_spectra_plot_WHEN_adding_more_traces_THEN_no_exceptions_thrown(self):
         g.begin()
         try:
@@ -63,7 +64,35 @@ class TestPlotting(unittest.TestCase):
         finally:
             g.end()
 
+    @retry_on_failure(3)
     def test_GIVEN_when_plot_exists_WHEN_connect_to_matplotlib_server_THEN_response_is_http_200_ok(self):
         TestPlotting.PYPLOT.plot(range(5))
         TestPlotting.PYPLOT.show()
         self.assert_webserver_up()
+
+    @retry_on_failure(3)
+    def test_WHEN_seven_spectra_added_THEN_six_figures_exist(self):
+        g.begin()
+        import matplotlib.pyplot as p
+        for i in range(7):
+            g.plot_spectrum(1)
+
+        self.assertEqual(6, len(p.get_fignums()))
+
+    @retry_on_failure(3)
+    def test_WHEN_seven_figures_added_THEN_six_figures_exist(self):
+        g.begin()
+        import matplotlib.pyplot as p
+        for i in range(7):
+            p.figure(i)
+
+        self.assertEqual(6, len(p.get_fignums()))
+
+    @retry_on_failure(3)
+    def test_WHEN_eight_reverse_numbered_figures_added_THEN_newest_six_exist(self):
+        g.begin()
+        import matplotlib.pyplot as p
+        for i in range(8, 0, -1):
+            p.figure(i)
+
+        self.assertEqual([1, 2, 3, 4, 5, 6], p.get_fignums())
